Faster recallc, Faster already seen unknown morphs, add ProperNoun list
I made recallc faster by using add_notes and add_cards, now it's able to recalculate 40k cards in under 2 minutes.  I made already seen be so much faster that it no longer freezes anki, the way I did that is it no longer checks every card for unknown morphs only cards that the user has specified to use with morphman, also now it only looks for unknown morphs in cads that have been reviewed for the first time on the same day because if a card has already been reviewed it can't have recent unknown morphs. Also added the ability for users to add lists of proper nouns in their porifiel folder